+++ CHANGES.rst
@@ -120,6 +120,8 @@ Changelog
 - Extend datagrid widget for doclink to select and display document [quang]
 - Remove getSelectionList from doclink [quang]
 - Add option to not load data on search form on intial display [quang]
+- Add 'Allow other value' option to selection fields to alow add new option on-the-fly [quang]
+- Add 'Unique field values' macro to select unique values stored of given document field [quang]
 
 2.0b4 (2016-05-27)
 ------------------